# ARROW-13266: [JS] Improve benchmark names & include suite name in json