--- a/js/perf/index.ts
+++ b/js/perf/index.ts
@@ -46,6 +46,7 @@ const results: CaseResult[] = [];
 function cycle(result: CaseResult, _summary: Summary) {
     const duration = result.details.median * 1000;
     if (json) {
+        result.suite = _summary.name
         results.push(result);
     }
     console.log(