Before:

```
Prepare Data: 635.29ms
Running "Parse "tracks"" suite...
Table.from 6,513 ops/s ±23%, 0.12 ms, 75 samples
readBatches 6,689 ops/s ±9.0%, 0.12 ms, 70 samples
serialize 325 ops/s ±11%, 2.7 ms, 69 samples
Running "Get "tracks" values by index" suite...
name: 'lat', length: 1,000,000, type: Float32 27.4 ops/s ±8.1%, 33 ms, 49 samples
name: 'lng', length: 1,000,000, type: Float32 28.5 ops/s ±3.8%, 33 ms, 50 samples
name: 'origin', length: 1,000,000, type: Dictionary<Int8, Utf8> 0.181 ops/s ±27%, 5,070 ms, 5 samples
name: 'destination', length: 1,000,000, type: Dictionary<Int8, Utf8> 0.175 ops/s ±28%, 5,619 ms, 5 samples
Running "Iterate "tracks" vectors" suite...
name: 'lat', length: 1,000,000, type: Float32 46 ops/s ±9.8%, 19 ms, 63 samples
name: 'lng', length: 1,000,000, type: Float32 52 ops/s ±1.6%, 19 ms, 67 samples
name: 'origin', length: 1,000,000, type: Dictionary<Int8, Utf8> 0.255 ops/s ±2.7%, 3,903 ms, 5 samples
name: 'destination', length: 1,000,000, type: Dictionary<Int8, Utf8> 0.225 ops/s ±14%, 4,328 ms, 5 samples
Running "Slice toArray "tracks" vectors" suite...
name: 'lat', length: 1,000,000, type: Float32 1,069 ops/s ±7.6%, 0.84 ms, 75 samples
name: 'lng', length: 1,000,000, type: Float32 892 ops/s ±7.3%, 1.1 ms, 61 samples
name: 'origin', length: 1,000,000, type: Dictionary<Int8, Utf8> 0.151 ops/s ±21%, 6,730 ms, 5 samples
name: 'destination', length: 1,000,000, type: Dictionary<Int8, Utf8> 0.189 ops/s ±24%, 5,173 ms, 5 samples
Running "Slice "tracks" vectors" suite...
name: 'lat', length: 1,000,000, type: Float32 1,815,763 ops/s ±4.9%, 0.001 ms, 86 samples
name: 'lng', length: 1,000,000, type: Float32 1,355,757 ops/s ±8.3%, 0.001 ms, 63 samples
name: 'origin', length: 1,000,000, type: Dictionary<Int8, Utf8> 1,678,921 ops/s ±15%, 0 ms, 71 samples
name: 'destination', length: 1,000,000, type: Dictionary<Int8, Utf8> 2,214,331 ops/s ±3.0%, 0 ms, 85 samples
Running "DataFrame Iterate "tracks"" suite...
length: 1,000,000 19.6 ops/s ±8.8%, 46 ms, 36 samples
Running "DataFrame Direct Count "tracks"" suite...
name: 'lat', length: 1,000,000, type: Float32, test: gt, value: 0 113 ops/s ±5.3%, 8.2 ms, 72 samples
name: 'lng', length: 1,000,000, type: Float32, test: gt, value: 0 89 ops/s ±13%, 9 ms, 59 samples
name: 'origin', length: 1,000,000, type: Dictionary<Int8, Utf8>, test: eq, value: Seattle 0.304 ops/s ±9.1%, 3,305 ms, 5 samples
```

After:

```
Prepare Data: 4.638ms
Running "Parse" suite...
dataset: tracks, function: Table.from 22,176 ops/s ±16%, 0.037 ms, 84 samples
dataset: tracks, function: readBatches 23,428 ops/s ±5.0%, 0.038 ms, 80 samples
dataset: tracks, function: serialize 14,132 ops/s ±5.8%, 0.063 ms, 74 samples
Running "Get values by index" suite...
dataset: tracks, column: lat, length: 200, type: Float32 450,042 ops/s ±5.9%, 0.002 ms, 91 samples
dataset: tracks, column: lng, length: 200, type: Float32 411,900 ops/s ±6.1%, 0.002 ms, 78 samples
dataset: tracks, column: origin, length: 200, type: Dictionary<Int8, Utf8> 991 ops/s ±11%, 0.96 ms, 60 samples
dataset: tracks, column: destination, length: 200, type: Dictionary<Int8, Utf8> 1,171 ops/s ±7.2%, 0.72 ms, 70 samples
Running "Iterate vectors" suite...
dataset: tracks, column: lat, length: 200, type: Float32 225,008 ops/s ±3.6%, 0.004 ms, 89 samples
dataset: tracks, column: lng, length: 200, type: Float32 240,921 ops/s ±0.51%, 0.004 ms, 97 samples
dataset: tracks, column: origin, length: 200, type: Dictionary<Int8, Utf8> 1,161 ops/s ±5.5%, 0.78 ms, 65 samples
dataset: tracks, column: destination, length: 200, type: Dictionary<Int8, Utf8> 1,088 ops/s ±5.6%, 0.96 ms, 42 samples
Running "Slice toArray vectors" suite...
dataset: tracks, column: lat, length: 200, type: Float32 651,617 ops/s ±3.8%, 0.002 ms, 83 samples
dataset: tracks, column: lng, length: 200, type: Float32 673,205 ops/s ±3.0%, 0.001 ms, 63 samples
dataset: tracks, column: origin, length: 200, type: Dictionary<Int8, Utf8> 1,145 ops/s ±5.7%, 0.77 ms, 69 samples
dataset: tracks, column: destination, length: 200, type: Dictionary<Int8, Utf8> 1,142 ops/s ±5.6%, 0.77 ms, 65 samples
Running "Slice vectors" suite...
dataset: tracks, column: lat, length: 200, type: Float32 5,699,749 ops/s ±1.8%, 0 ms, 93 samples
dataset: tracks, column: lng, length: 200, type: Float32 5,768,406 ops/s ±0.75%, 0 ms, 97 samples
dataset: tracks, column: origin, length: 200, type: Dictionary<Int8, Utf8> 6,112,340 ops/s ±0.49%, 0 ms, 96 samples
dataset: tracks, column: destination, length: 200, type: Dictionary<Int8, Utf8> 6,033,494 ops/s ±1.3%, 0 ms, 94 samples
Running "DataFrame Iterate" suite...
dataset: tracks, length: 200 106,752 ops/s ±3.1%, 0.009 ms, 86 samples
Running "DataFrame Direct Count" suite...
dataset: tracks, column: lat, length: 200, type: Float32, test: gt, value: 0 1,277,478 ops/s ±2.3%, 0.001 ms, 83 samples
dataset: tracks, column: lng, length: 200, type: Float32, test: gt, value: 0 1,163,103 ops/s ±3.9%, 0.001 ms, 74 samples
dataset: tracks, column: origin, length: 200, type: Dictionary<Int8, Utf8>, test: eq, value: Seattle 1,209 ops/s ±7.1%, 0.76 ms, 63 samples
```